requirements.txt
wheel==0.37.1
asgiref==3.4.1
Django==3.2.6
mysqlclient==2.0.1
pytz==2021.1
sqlparse==0.4.1
django-debug-toolbar==2.2
django-rest-authtoken==2.1.4
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
django-allauth==0.47.0
requests==2.26.0
boto3==1.20.28
django-storages==1.12.3
Pillow==9.0.0
django-extensions==3.1.5 